Refactored a method into shared util class and documented it.
