+++ http-builder-ng-core/src/main/java/groovyx/net/http/util/IoUtils.java
@@ -15,10 +15,13 @@
  */
 package groovyx.net.http.util;
 
+import groovyx.net.http.TransportingException;
+
 import java.io.BufferedInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 
 /**
  * Shared IO utility operations.
@@ -52,9 +55,17 @@ public class IoUtils {
         }
     }
 
-    // FIXME: document
+    /**
+     * Safely copies the contents of the {@link BufferedInputStream} to a {@link String} and resets the stream.
+     * This method is generally only useful for testing and logging purposes.
+     *
+     * @param inputStream the BufferedInputStream
+     * @return a String copy of the stream contents
+     * @throws IOException           if something goes wrong with the stream
+     * @throws IllegalStateException if the stream cannot be reset
+     */
     public static String copyAsString(final BufferedInputStream inputStream) throws IOException, IllegalStateException {
-        if (inputStream == null ) return null;
+        if (inputStream == null) return null;
 
         try {
             inputStream.mark(Integer.MAX_VALUE);
@@ -64,6 +75,33 @@ public class IoUtils {
                 inputStream.reset();
             } catch (IOException ioe) {
                 throw new IllegalStateException("Unable to reset stream - original stream content may be corrupted");
+            }
+        }
+    }
+
+    /**
+     * Transfers the contents of the {@link InputStream} into the {@link OutputStream}, optionally closing the stream.
+     *
+     * @param istream the input stream
+     * @param ostream the output stream
+     * @param close   whether or not to close the output stream
+     */
+    public static void transfer(final InputStream istream, final OutputStream ostream, final boolean close) {
+        try {
+            final byte[] bytes = new byte[2_048];
+            int read;
+            while ((read = istream.read(bytes)) != -1) {
+                ostream.write(bytes, 0, read);
+            }
+        } catch (IOException e) {
+            throw new TransportingException(e);
+        } finally {
+            if (close) {
+                try {
+                    ostream.close();
+                } catch (IOException ioe) {
+                    throw new TransportingException(ioe);
+                }
             }
         }
     }
